Remove /etc/initrd-release from live-files
We added this because that file was missing from MX and
we wanted to copy the linxuxfs version to the new initrd.
We should be able to ensure it is in all productions iso
files going forward so it is better to use the one from
the linuxfs file and not the one from this package.